Current Conditions Component Built
Parent component with sub-components, unit conversion fn moved to util
