+++ src/components/blocks/CampusWeatherDetail/CampusWeatherCurrentConditions/CampusWeatherCurrentConditions.module.scss
@@ -1,0 +1,23 @@
+.CurrentConditionsContainer {
+	flex: 1 1;
+	display: flex;
+	flex-wrap: wrap;
+	gap: 10px;
+
+	.weatherSymbol {
+		flex: 1 1 100%;
+		text-align: center;
+		border-radius: 50%;
+		margin: 0 auto;
+		padding: 50px;
+	}
+	.weatherDescription {
+		flex: 1 1 100%;
+		text-align: center;
+	}
+	.airTemperature {
+		flex: 1 1 100%;
+		text-align: center;
+		font-size: 2em;
+	}
+}
