+++ src/components/blocks/CampusWeatherDetail/CampusWeatherCurrentConditions/CampusWeatherCurrentValue/CampusWeatherCurrentValue.module.scss
@@ -1,0 +1,10 @@
+.parameter {
+	flex: 1 1;
+	text-align: center;
+	.label {
+		font-size: 0.8em;
+	}
+	.value {
+		font-size: 1em;
+	}
+}
