+++ src/components/blocks/CampusWeatherDetail/CampusWeatherDetail.module.scss
@@ -7,44 +7,6 @@
 		flex: 1 1 100%;
 		text-align: center;
 	}
-
-	.CurrentConditionsContainer {
-		flex: 1 1;
-		display: flex;
-		flex-wrap: wrap;
-		gap: 10px;
-
-		.weatherSymbol {
-			flex: 1 1 100%;
-			text-align: center;
-			border-radius: 50%;
-			margin: 0 auto;
-			padding: 50px;
-		}
-		.weatherDescription {
-			flex: 1 1 100%;
-			text-align: center;
-		}
-		.airTemperature {
-			flex: 1 1 100%;
-			text-align: center;
-			font-size: 2em;
-		}
-
-		.apparentTemperature,
-		.dewpointTemperature,
-		.humidity,
-		.wind {
-			flex: 1 1;
-			text-align: center;
-			.label {
-				font-size: 0.8em;
-			}
-			.value {
-				font-size: 1em;
-			}
-		}
-	}
 	.Spacer {
 		flex: 2 1 auto; /* Takes up remaining space */
 	}
